{CODEC2] Converted all Codec2 inputs/outputs to floating point.

+++ DSM_D407/Src/Codec2/src/codec2.h
@@ -48,8 +48,8 @@ extern "C" {
 void codec2_init(struct CODEC2 *codec2_state, int mode);
 void codec2_close(struct CODEC2 *codec2_state);
 
-void codec2_encode(struct CODEC2 *codec2_state, unsigned char * bits, short speech_in[]);
-void codec2_decode(struct CODEC2 *codec2_state, short speech_out[], const unsigned char *bits);
+void codec2_encode(struct CODEC2 *codec2_state, unsigned char * bits, float speech_in[]);
+void codec2_decode(struct CODEC2 *codec2_state, float speech_out[], const unsigned char *bits);
 
 int  codec2_state_memory_req(void);
 int  codec2_samples_per_frame(struct CODEC2 *codec2_state);
